If --term_existing argument is passed then any existing mercurywallet.exe processes will be terminated (windows version)

+++ public/electron.js
@@ -114,14 +114,14 @@ async function getTorAdapter(path) {
 // initialization and is ready to create browser windows.
 // Some APIs can only be used after this event occurs.
 app.on('ready', () => {
-  //Limit app to one instance
-  if(!app.requestSingleInstanceLock() && getPlatform() !== 'win'){
+    //Limit app to one instance
+    if(!app.requestSingleInstanceLock() && !(term_existing && getPlatform() === 'win') ){
       alert('mercurywallet is already running. Not opening app.')
       app.quit()
-    } 
+    }
   
-  terminate_mercurywallet_process(init_tor_adapter);
-  createWindow()
+    terminate_mercurywallet_process(init_tor_adapter);
+    createWindow()
   }
 );
 
@@ -311,8 +311,10 @@ async function on_exit(){
 }
 
 async function kill_tor(){
-  console.log("terminating the tor adapter process...")
-  await kill_process(ta_process.pid)
+  if(ta_process){
+    console.log("terminating the tor adapter process...")
+    await kill_process(ta_process.pid)
+  }
 }
 
 async function kill_process(pid, init_new){
